Major refactor. Cleaned up the bartleby root directory by moving classes and functions to their own directories. Split helper_functions up by adding seperate modules for model prompting functions and command parsing functions. Move listener and generator loop functions out of bartleby.py and into their own module. I think that's it.
